backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic==2.11.7
SQLAlchemy==2.0.41
psycopg[binary]==3.2.9
python-dotenv==1.0.1
google-auth==2.40.3
itsdangerous==2.2.0
requests==2.32.3
